+++ app/Seggu.Domain/Origin.cs
@@ -1,8 +1,6 @@
 namespace Seggu.Domain
-{
-    using System;
-    
-    public enum Origin : short
+{    
+    public enum Origin : int
     {
         National = 1,
         Imported = 2
